requirements.txt
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.14
channels==4.0.0
channels-redis==4.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-phonenumber-field==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
msgpack==1.1.1
phonenumbers==8.13.27
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-telegram-bot==20.7
redis==6.2.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
